-experiencetolevel type fix

+++ LineageTools.Experience.DB/Models/ExperienceToLevel.cs
@@ -5,6 +5,6 @@ namespace LineageTools.Experience.DB.Models
 	public class ExperienceToLevel : BaseEntity
 	{
 		public int Level { get; set; }
-		public int ExpCount { get; set; }
+		public long ExpCount { get; set; }
 	}
 }
